requirements.txt
fastapi
uvicorn
gunicorn
motor
python-dotenv